+++ apps/openmw/mwmechanics/mechanicsmanagerimp.cpp
@@ -1480,7 +1480,7 @@ namespace MWMechanics
             disposition = getDerivedDisposition(ptr);
 
         int fight = std::max(0, ptr.getClass().getCreatureStats(ptr).getAiSetting(CreatureStats::AI_Fight).getModified()
-                + static_cast<int>(getFightDistanceBias(ptr, target) + getFightDispositionBias(disposition)));
+                + static_cast<int>(getFightDistanceBias(ptr, target) + getFightDispositionBias(static_cast<float>(disposition))));
 
         if (ptr.getClass().isNpc() && target.getClass().isNpc())
         {
